Improve content loading checks
The `await` we had in `main.js:startup` was there to block the loading of content until after
critical resrouces like mathjax and elkjs had been loaded; however, this safeguard applies
only to content being re-loaded, according to the stored state. While it's unlikely the user
could manually attempt to load any content before the sibling scripts had had time to finish
loading, it would be nice to have a truly universal safeguard. This commit does that.

Now, `main.js:startup` goes back to being a plain, synchronous function, and we instead make
the `ContentManager` check with the `Hub` to see that the sibling scripts have finished loading,
before it goes ahead and loads any content.